% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_values.R
\name{color_values}
\alias{color_values}
\title{color values
maps colours to values}
\usage{
color_values(x, palette = "viridis", na_colour = "#808080")
}
\arguments{
\item{x}{vector of values to map to a colour}

\item{palette}{palette to use}

\item{na_colour}{hex string colour to use for NA values}
}
\description{
color values
maps colours to values
}
\seealso{
colour_values
}
