% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_values.R
\name{color_values_rgb}
\alias{color_values_rgb}
\title{Colour Values RGB}
\usage{
color_values_rgb(x, palette = "viridis", na_colour = "#808080FF",
  alpha = 255, include_alpha = TRUE)
}
\description{
maps colours to values, returning a matrix of RGB(A) values
}
\seealso{
colour_values_rgb
}
