% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_values.R
\name{color_values}
\alias{color_values}
\title{color values
maps colours to values}
\usage{
color_values(x, palette = "viridis", na_colour = "#808080FF", alpha = 255,
  return = c("hex", "rgb"))
}
\arguments{
\item{x}{vector of values to map to a colour}

\item{palette}{colour palette. See details and examples}

\item{na_colour}{hex string colour to use for NA values in the form \code{#RRGGBBAA}.}

\item{alpha}{optional. Single value in [0,255] applied to all colours, or
a vector of numeric values the same length as \code{x}. The numeric vector
will be scaled into the range [0,255].
If a matrix palette is supplied this argument is ignored.}

\item{return}{either 'hex' or 'rgb'. If 'hex' hex colours are returned.
If 'rgb' a length(\code{x}) x 4-column matrix is returned with columns in
the order red, green, blue and alpha, and each row corresponds to each elemetn of \code{x}}
}
\description{
color values
maps colours to values
}
\section{Functions}{
\itemize{
\item \code{color_values}: Values
}}

\seealso{
colour_values
}
