% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{colour_palettes}
\alias{colour_palettes}
\alias{color_palettes}
\title{Colour Palettes}
\usage{
colour_palettes()

color_palettes()
}
\description{
List the available colour palettes.
}
\examples{
colour_palettes()

}
