% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_values.R
\name{colour_values}
\alias{colour_values}
\title{Colour Values}
\usage{
colour_values(x, palette = "viridis", na_colour = "#808080")
}
\arguments{
\item{x}{vector of values to map to a colour}

\item{palette}{palette to use}

\item{na_colour}{hex string colour to use for NA values}
}
\description{
maps colours to values
}
\examples{

colour_values(x = 1:5) ## default is "viridis"
colour_values(x = 1:5, palette = "inferno")
colour_values(x = 1:5, palette = "plasma")
colour_values(x = 1:5, palette = "magma")
colour_values(x = 1:5, palette = "cividis")

}
