% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_values_rgb.R
\name{colour_values_rgb}
\alias{colour_values_rgb}
\title{Colour Values RGB}
\usage{
colour_values_rgb(x, palette = "viridis", alpha = 255,
  na_colour = "#808080FF", include_alpha = TRUE, ...)
}
\arguments{
\item{x}{vector of values to map to a colour}

\item{palette}{colour palette. See details and examples}

\item{alpha}{optional. Single value in [0,255] applied to all colours,
or a decimal in [0, 1) (to indicate a percentage, noting 1 is excluded),
or a vector of numeric values the same length as \code{x}.
The numeric vector will be scaled into the range [0,255].
If a matrix palette is supplied this argument is ignored.}

\item{na_colour}{hex string colour to use for NA values in the form \code{#RRGGBBAA}.}

\item{include_alpha}{logical indicating if the returned hex or matrix should include
the alpha values. Defaults to \code{TRUE}.}

\item{...}{other arguments possed to methods}
}
\description{
Maps colours to variables, returning a matrix of RGB(A) values
}
\details{
The palette can either be
\itemize{
  \item{String - }{use \code{colour_palettes()} to view available palettes}
  \item{Matrix - }{At least 5 rows, and 3 (or 4) columns representing the red, green and blue (and alpha) values}
}

The matrix \code{palette} requires 5 rows because the colours are interpolated using a
cubic b-spline. This method requires 5 values.
}
\examples{

colour_values_rgb(1:5)
colour_values_rgb(1:5, include_alpha = FALSE)
colour_values_rgb(-25:25, n_summaries = 5)


}
\seealso{
colour_values
}
