% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_variables.R
\name{colour_variables.default}
\alias{colour_variables.default}
\title{#' @export
colour_variables.Date <- function( x, palette = "viridis", na_colour = "#808080" ) {
  rcpp_colour_dte_variable_hex( x, palette, na_colour )
}}
\usage{
\method{colour_variables}{default}(x, palette = "viridis",
  na_colour = "#808080")
}
\description{
#' @export
colour_variables.Date <- function( x, palette = "viridis", na_colour = "#808080" ) {
  rcpp_colour_dte_variable_hex( x, palette, na_colour )
}
}
