% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_variables.R
\name{colour_variables}
\alias{colour_variables}
\title{Colour Variables}
\usage{
colour_variables(x, palette = c("viridis", "inferno", "plasma", "magma",
  "cividis"))
}
\arguments{
\item{x}{vector of values to map to a colour}

\item{palette}{palette to use}
}
\description{
maps a colours to variables
}
\examples{

colour_variables(x = 1:5)
colour_variables(x = 1:5, palette = "inferno")
colour_variables(x = 1:5, palette = "plasma")
colour_variables(x = 1:5, palette = "magma")
colour_variables(x = 1:5, palette = "cividis")

}
