% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_colours.R
\name{convert_color}
\alias{convert_color}
\title{Convert Color}
\usage{
convert_color(x)
}
\arguments{
\item{x}{character vector of hex strings, or numeric matrix of RRGGBBAA values}
}
\description{
Converts colours between RRGGBBAA and hex strings, in both directions.
}
\seealso{
convert_colour
}
