% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_colours.R
\name{convert_colors}
\alias{convert_colors}
\title{Convert Colors}
\usage{
convert_colors(x)
}
\arguments{
\item{x}{character vector of hex strings, or numeric matrix of RRGGBBAA values}
}
\description{
Converts colours between RRGGBBAA and hex strings, in both directions.
}
\seealso{
convert_colour
}
