% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_colours.R
\name{convert_colour}
\alias{convert_colour}
\title{Convert}
\usage{
convert_colour(x)
}
\arguments{
\item{x}{character vector of hex strings, or numeric matrix of RRGGBBAA values}
}
\description{
Converts colours between RRGGBBAA and hex strings, in both directions
}
\details{
If a combination of hex strings with and without alpha values are supplied, those
without are assumed to have an alpha value of \code{FF} and will be returned in the
RRGGBBAA matrix
}
\examples{

convert_colour(c("#FFAA00"))


}
