% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_variables.R
\name{colour_variables}
\alias{colour_variables}
\title{Colour Variables}
\usage{
colour_variables(x, palette = "viridis", na_colour = "#808080")
}
\arguments{
\item{x}{vector of values to map to a colour}

\item{palette}{palette to use}

\item{na_colour}{hex string}
}
\description{
maps a colours to variables
}
\examples{

colour_variables(x = 1:5) ## default is "viridis"
colour_variables(x = 1:5, palette = "inferno")
colour_variables(x = 1:5, palette = "plasma")
colour_variables(x = 1:5, palette = "magma")
colour_variables(x = 1:5, palette = "cividis")

}
