% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_palette}
\alias{get_palette}
\title{Get Palette}
\usage{
get_palette(palette)
}
\arguments{
\item{palette}{one of the available paletes. See \link{colour_palettes}}
}
\value{
256 row x 3 column matrix. Columns are in the order red, green, blue.
}
\description{
retrieves one of the available palettes
}
\examples{

get_palette( "viridis" )
get_palette( "rainbow" )

}
