% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_colours}
\alias{show_colours}
\title{Show Colours}
\usage{
show_colours(colours = colour_palettes())
}
\arguments{
\item{colours}{vector of colour palettes}
}
\description{
Plots all the selected colours. See \link{colour_palettes} for avaialble colours.
}
\examples{

## view all the colour palettes
show_colours()

## view a selection of colour palettes
show_colours( colours = colour_palettes( c("viridis", "grdevices") ) )



}
